# crates/server/openapi.json
{
  "components": {
    "schemas": {
...
      "EventName": {
        "description": "The one name the app's allow-list accepts today (`docs/v1/server.md`, H13); any other value\nfails to deserialize, and answers `request.malformed`.",
        "enum": [
          "export_completed"
        ],
        "type": "string"
      },
      "EventRequest": {
        "description": "A product event the app reports, with the browsing context every one of them carries.",
        "properties": {
          "appVersion": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/EventName"
          },
          "platform": {
            "type": "string"
          },
          "properties": {
            "$ref": "#/components/schemas/ExportCompletedProperties"
          }
        },
        "required": [
          "name",
          "properties",
          "platform",
          "appVersion",
          "language"
        ],
        "type": "object"
      },
      "ExportCompletedProperties": {
        "description": "`export_completed`'s properties: the format downloaded and its raw, uncompressed size.",
        "properties": {
          "bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          }
        },
        "required": [
          "format",
          "bytes"
        ],
        "type": "object"
      },
...
      "ExportFormat": {
        "description": "The format a completed export used, as `docs/v1/service.md` fixes the list.",
        "enum": [
          "wasm",
          "gif",
          "apng",
          "sprite_sheet",
          "png_frames"
        ],
        "type": "string"
      },
...
      "Problem": {
        "description": "An error, as RFC 9457 problem details with a stable code and its parameters, never a\nsentence: the interface translates `errors.<code>`.",
        "example": {
          "code": "quota.storage_exceeded",
          "params": {
            "limit": 100000000,
            "requested": 2000000,
            "used": 99000000
          },
          "status": 409,
          "type": "urn:life-pixel:problem:quota.storage_exceeded"
        },
        "properties": {
          "code": {
            "description": "The stable code, dot-separated `snake_case` segments, domain first.",
            "type": "string"
          },
          "params": {
            "additionalProperties": {},
            "description": "The parameters of the code's message, with camelCase keys.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "status": {
            "description": "The HTTP status.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "type": {
            "description": "`urn:life-pixel:problem:<code>`.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "code",
          "params"
        ],
        "type": "object"
      },
...
      "SupportAuthor": {
        "description": "Who wrote a message.",
        "enum": [
          "user",
          "team"
        ],
        "type": "string"
      },
      "SupportCategory": {
        "description": "What a support request is about.",
        "enum": [
          "bug",
          "account",
          "billing",
          "data_protection",
          "abuse",
          "other"
        ],
        "type": "string"
      },
...
      "SupportRequestMessage": {
        "description": "A message of a request: never an internal note of the team.",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/SupportAuthor",
            "description": "Who wrote it."
          },
          "body": {
            "description": "Its text.",
            "type": "string"
          },
          "createdAt": {
            "description": "When it was written.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Its id.",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author",
          "body",
          "createdAt"
        ],
        "type": "object"
      },
      "SupportRequestStatus": {
        "description": "Where a support request stands.",
        "enum": [
          "new",
          "in_progress",
          "waiting_for_user",
          "resolved",
          "closed"
        ],
        "type": "string"
      },
...
      "SupportRequestThread": {
        "description": "A request of the account and its messages, oldest first: the first is the request's own.",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/SupportCategory",
            "description": "What it is about."
          },
          "createdAt": {
            "description": "When it was sent.",
            "format": "date-time",
            "type": "string"
          },
          "hasScreenshot": {
            "description": "Whether a screenshot came with it; the screenshot itself is never served back.",
            "type": "boolean"
          },
          "id": {
            "description": "Its id.",
            "format": "uuid",
            "type": "string"
          },
          "messages": {
            "description": "Its messages, the team's internal notes left out.",
            "items": {
              "$ref": "#/components/schemas/SupportRequestMessage"
            },
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/SupportRequestStatus",
            "description": "Where it stands."
          },
          "updatedAt": {
            "description": "When it last changed.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "category",
          "status",
          "hasScreenshot",
          "createdAt",
          "updatedAt",
          "messages"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "csrf": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "session": {
        "in": "cookie",
        "name": "__Host-lp_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The Life Pixel server's public API.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "Life Pixel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "operationId": "createEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Recorded, or silently dropped under load"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`request.malformed`: a name, property or value outside the allow-list"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`rate_limit.exceeded`: 60 a minute per address"
          }
        },
        "summary": "Records an allowed product event; the subject comes from the session when there is one. Once\nthe request parses, this never fails: a full channel only drops the event, counted by\n`events_dropped_total`.",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/support-requests/{id}": {
      "get": {
        "operationId": "getSupportRequest",
        "parameters": [
          {
            "description": "The request's id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportRequestThread"
                }
              }
            },
            "description": "The request and its messages"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`request.malformed`: a cursor, an id, a query or a body that does not parse"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`auth.unauthenticated`: no session, or one that ended"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`auth.account_suspended`"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`support.request_not_found`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`rate_limit.exceeded`: 600 a minute per account"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "The request `id`, with its messages: the team's internal notes never appear, and neither\ndoes its screenshot.",
        "tags": [
          "support"
        ]
      }
    },
...
  }
}